make MPD::Connection::Version() return integer instead of float
full version is always 0.xx.0, where xx is current
version of protocol, so why bother with zeros...

+++ src/mpdpp.h
@@ -92,7 +92,7 @@ namespace MPD
 			const std::string & GetHostname() { return itsHost; }
 			int GetPort() { return itsPort; }
 			
-			float Version() const;
+			unsigned Version() const;
 			
 			void SetIdleEnabled(bool val) { isIdleEnabled = val; }
 			bool SupportsIdle() const { return supportsIdle; }
